Condition for delete heroe

+++ src/app/heroes/pages/add-heroes/add-heroes.component.ts
@@ -5,6 +5,8 @@ import { switchMap } from "rxjs/operators";
 import { HeroesService } from '../../services/heroes.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmarComponent } from '../../components/confirmar/confirmar.component';
 @Component({
   selector: 'app-add-heroes',
   templateUrl: './add-heroes.component.html',
@@ -39,7 +41,8 @@ export class AddHeroesComponent implements OnInit {
   constructor(private heroesService: HeroesService,
               private activatedRoute: ActivatedRoute, 
               private router: Router,
-              private snackBar: MatSnackBar) { }
+              private snackBar: MatSnackBar,
+              public dialog: MatDialog) { }
 
   ngOnInit(): void {
 
@@ -72,10 +75,22 @@ export class AddHeroesComponent implements OnInit {
   }
 
   borrarHeroe(){
-    this.heroesService.deleteHeroe(this.heroe.id!)
-    .subscribe(resp => {
-      this.router.navigate(['/heroes']);
-    })
+    const dialog = this.dialog.open(ConfirmarComponent, {
+      width: '250px',
+      data: {...this.heroe}
+    });
+
+    dialog.afterClosed().subscribe(
+      (result) => {
+        if (result){
+          this.heroesService.deleteHeroe(this.heroe.id!)
+          .subscribe(resp => {
+            this.router.navigate(['/heroes']);
+          })
+        }
+      }
+    )
+    
   }
 
   mostrarsnackBar(mensaje: string): void{
